+++ perl/FSM/Scheduler/ISF/LoweringIR.pm
@@ -17,7 +17,7 @@ my %SUPPORTED_TRANSACTION_CLAUSES = (
             on drive await sample update phase shift_left shift_right assemble
             extract complete when switch repeat latency do spawn await_all
             await_any params stage contract store load wait while until set
-            atl_trigger
+            atl_trigger atl_group_trigger
         )
     },
     when => {
@@ -374,6 +374,21 @@ sub _actor_network_for_ir {
             scheduling  => $_->{scheduling},
         }
     } @{$network->{groups} || []};
+    my @group_schedules = map {
+        {
+            group               => $_->{group},
+            owner_transaction   => $_->{owner_transaction},
+            context             => $_->{context},
+            members             => [ @{$_->{members} || []} ],
+            target_transactions => [ @{$_->{target_transactions} || []} ],
+            signals             => [ @{$_->{signals} || []} ],
+            schedule            => $_->{schedule},
+            dependency_policy   => $_->{dependency_policy},
+            storage             => $_->{storage},
+            source              => $_->{source},
+            sink                => $_->{sink},
+        }
+    } @{$network->{group_schedules} || []};
 
     return {
         kind      => $network->{kind} // 'static_declaration',
@@ -387,6 +402,7 @@ sub _actor_network_for_ir {
             } @{$network->{instances}}
         ],
         groups => \@groups,
+        group_schedules => \@group_schedules,
         event_waits => \@event_waits,
         transaction_triggers => \@transaction_triggers,
         data_movements => \@data_movements,
@@ -2593,6 +2609,7 @@ sub _build_transaction($self, $tx, $actor, $txi, $generated_children = undef) {
         }
         elsif ($k eq 'await')    { $ha=1; $wdc="${tn}_wd"; my $wd_override = _parse_await_wd($cl); push @st, _ir_await($cl, $tn, $si++, $wd_override || $wd, [splice @ps]); }
         elsif ($k eq 'atl_trigger') { push @st, _ir_atl_trigger($cl, $tn, $si++, [splice @ps]); }
+        elsif ($k eq 'atl_group_trigger') { push @st, _ir_atl_group_trigger($cl, $tn, $si++, [splice @ps]); }
         elsif ($k eq 'sample')   { push @ps, $cl; }
         elsif ($k eq 'wait') {
             my $wait = _wait_count_spec($cl, $tn, 'transaction body', $actor, $widths, 1);
@@ -4750,6 +4767,25 @@ sub _ir_atl_trigger {
 
     return {
         name        => "${tn}_atl_trigger_$i",
+        kind        => 'sequential',
+        assignments => \@assignments,
+        transitions => [],
+    };
+}
+sub _ir_atl_group_trigger {
+    my ($cl, $tn, $i, $pending_samples) = @_;
+    my @assignments = _sample_assignments($pending_samples || []);
+    for my $signal (@{$cl}[1 .. $#$cl]) {
+        push @assignments, {
+            lhs         => $signal,
+            rhs         => 1,
+            op          => '<1',
+            source_kind => 'atl_group_actor_transaction_trigger',
+        };
+    }
+
+    return {
+        name        => "${tn}_atl_group_trigger_$i",
         kind        => 'sequential',
         assignments => \@assignments,
         transitions => [],