+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -98,6 +98,7 @@ our @EXPORT_OK = qw(
     isf_public_interface_schedule_report_actor_stage_keys
     isf_public_interface_schedule_report_actor_network_data_movement_keys
     isf_public_interface_schedule_report_actor_network_event_wait_keys
+    isf_public_interface_schedule_report_actor_network_group_schedule_keys
     isf_public_interface_schedule_report_actor_network_group_keys
     isf_public_interface_schedule_report_actor_network_instance_keys
     isf_public_interface_schedule_report_actor_network_keys
@@ -224,6 +225,7 @@ sub build_isf_public_interface_contract {
         schedule_report_actor_network_keys => isf_public_interface_schedule_report_actor_network_keys(),
         schedule_report_actor_network_instance_keys => isf_public_interface_schedule_report_actor_network_instance_keys(),
         schedule_report_actor_network_group_keys => isf_public_interface_schedule_report_actor_network_group_keys(),
+        schedule_report_actor_network_group_schedule_keys => isf_public_interface_schedule_report_actor_network_group_schedule_keys(),
         schedule_report_actor_network_data_movement_keys => isf_public_interface_schedule_report_actor_network_data_movement_keys(),
         schedule_report_actor_network_event_wait_keys => isf_public_interface_schedule_report_actor_network_event_wait_keys(),
         schedule_report_actor_network_transaction_trigger_keys => isf_public_interface_schedule_report_actor_network_transaction_trigger_keys(),
@@ -575,6 +577,7 @@ sub isf_public_interface_public_top_level_keys {
             schedule_report_actor_network_keys
             schedule_report_actor_network_instance_keys
             schedule_report_actor_network_group_keys
+            schedule_report_actor_network_group_schedule_keys
             schedule_report_actor_network_data_movement_keys
             schedule_report_actor_network_event_wait_keys
             schedule_report_actor_network_transaction_trigger_keys
@@ -1317,6 +1320,7 @@ sub isf_public_interface_schedule_report_actor_network_keys {
             data_movements
             event_waits
             groups
+            group_schedules
             kind
             instances
             transaction_triggers
@@ -1343,6 +1347,24 @@ sub isf_public_interface_schedule_report_actor_network_group_keys {
             name
             scheduling
             source
+        ),
+    ];
+}
+
+sub isf_public_interface_schedule_report_actor_network_group_schedule_keys {
+    return [
+        qw(
+            context
+            dependency_policy
+            group
+            members
+            owner_transaction
+            schedule
+            signals
+            sink
+            source
+            storage
+            target_transactions
         ),
     ];
 }
@@ -1739,6 +1761,7 @@ sub isf_public_interface_schedule_report_presence_key_family_map {
         schedule_report_actor_network_keys => isf_public_interface_schedule_report_actor_network_keys(),
         schedule_report_actor_network_instance_keys => isf_public_interface_schedule_report_actor_network_instance_keys(),
         schedule_report_actor_network_group_keys => isf_public_interface_schedule_report_actor_network_group_keys(),
+        schedule_report_actor_network_group_schedule_keys => isf_public_interface_schedule_report_actor_network_group_schedule_keys(),
         schedule_report_actor_network_data_movement_keys => isf_public_interface_schedule_report_actor_network_data_movement_keys(),
         schedule_report_actor_network_event_wait_keys => isf_public_interface_schedule_report_actor_network_event_wait_keys(),
         schedule_report_actor_network_transaction_trigger_keys => isf_public_interface_schedule_report_actor_network_transaction_trigger_keys(),